Moved operation functions to a different file.
